Revert "CXO-11443 Merged with master. Moved black token service validation logic to TokenService."
This reverts commit 240afa11

+++ comu/com.energyict.mdc.engine.config/src/main/java/com/energyict/mdc/engine/config/ComServerAliveStatus.java
@@ -17,7 +17,7 @@ public interface ComServerAliveStatus {
 
     Instant getLastActiveTime();
 
-    void update(Instant time, Integer updateFrequency, Instant blockedSince, Long blockTime);
+    void update(Instant time, Integer updateFrequency, Instant blockedSince, Integer blockTime);
 
     boolean isBlocked();
 
@@ -29,6 +29,6 @@ public interface ComServerAliveStatus {
 
     Optional<Instant> getBlockedSince();
 
-    Optional<Long> getBlockedTime();
+    Optional<Integer> getBlockedTime();
 
 }
